refactor: play_pageのウィジェットをplay_flowに切り出し
