+++ lib/features/zazen/play_flow/zazen_ending.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:xen2/components/outlined_text.dart';
+
+class ZazenEnding extends StatelessWidget {
+  const ZazenEnding({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const OutlinedText(text: '終了', fontSize: 20);
+  }
+}
