+++ lib/features/zazen/play_flow/zazen_in_progress.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ZazenInProgress extends StatelessWidget {
+  const ZazenInProgress({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox.shrink();
+  }
+}
